// src/models/Producto.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface Producto {
  id: number;
  clave_prod: string;
  descripcion: string;
  descripcion_corta: string;
  sucursal_origen: number;
  fecha_alta: string;
  fecha_act: string;
  idMarca: number;
  marca: string;
  area: number;
  d_area: string;
  depto: number;
  d_depto: string;
  clase: number;
  d_clase: string;
  observacion: string;
  inventariable: boolean;
  controlado: boolean;
  es_fraccion: boolean;
  obsoleto: boolean;
  es_insumo: boolean;
  es_servicio: boolean;
  es_producto: boolean;
  es_kit: boolean;
  tasa_iva: number;
  tasa_ieps: number;
  costo_unitario: number;
  precio: number;
  unidad_paq: number;
  unidad_paq_traspaso: number;
  promocion: boolean;
  porcentaje_promocion: number;
  precio_promocion: number;
  fecha_inicio: string;
  fecha_final: string;
  unidad_medida: number;
  d_unidadMedida: string;
  clave_prov: number;
  d_proveedor: string;
  tiempo: number;
  comision: number;
  productoLibre: boolean;
}

// Generated by https://quicktype.io

export interface ProductoExistencia {
  id: number;
  clave_prod: string;
  descripcion: string;
  descripcion_corta: string;
  existencia: number;
  sucursal_origen: number;
  fecha_alta: string;
  fecha_act: string;
  idMarca: number;
  marca: Marca;
  area: number;
  d_area: DArea;
  depto: number;
  d_depto: DDepto;
  clase: number;
  d_clase: DClase;
  observacion: string;
  inventariable: boolean;
  controlado: boolean;
  es_fraccion: boolean;
  obsoleto: boolean;
  es_insumo: boolean;
  es_servicio: boolean;
  es_producto: boolean;
  es_kit: boolean;
  tasa_iva: number;
  tasa_ieps: number;
  costo_unitario: number;
  precio: number;
  unidad_paq: number;
  unidad_paq_traspaso: number;
  promocion: boolean;
  porcentaje_promocion: number;
  precio_promocion: number;
  fecha_inicio: null;
  fecha_final: null;
  unidad_medida: number;
  d_unidadMedida: DUnidadMedida;
  clave_prov: number;
  d_proveedor: null;
  tiempo: number;
  comision: number;
  productoLibre: boolean;
}

export enum DArea {
  Servicios120 = "Servicios 120",
}

export enum DClase {
  Descripcion = "Descripcion",
}

export enum DDepto {
  LuisPrueba221 = "LUIS PRUEBA221",
}

export enum DUnidadMedida {
  Pza = "Pza",
}

export enum Marca {
  NewBlack = "NEW BLACK",
}